pyproject.toml
[tool.poetry]
name = "tiltedbackendpy"
version = "0.1.0"
description = ""
authors = ["Ega2901"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
mysql-connector-python = "^9.0.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
